Aggregates the locally derived summary data using the proposed method SHIR.

Aggregates the locally derived summary data using the proposed method SHIR.

SHIR_fit(
  H_lst,
  d_lst,
  n_lst,
  lambda_lst = NULL,
  lambda_g_lst = NULL,
  tune = "BIC"
)

Arguments

H_lst
List of locally derived Hessian matrices

d_lst
List of locally derived gradient vectors

n_lst
Vector containing the sample size of each local site

lambda_lst
Candidate set of the tuning parameters for mu. It corresponds to sqrt(N)lambda in the SHIR paper. If not specified or specified as NULL by the user, default value is 0.3*c(5:25)sqrt(nlog(p)).

lambda_g_lst
Candidate set of the tuning parameter for alpha. It corresponds to lambda_g in the SHIR paper. If not specified or specified as NULL by the user, default value is c(0.6, 0.9).

tune
Information criterion used for model selection. Input options include 'AIC', 'BIC', 'mBIC' and 'RIC', which put different weights on the degree of freedom. If not specified by the user, default value is 'BIC'.

Value

min.lambda: Selected tuning parameters with the minimum information criterion.

min.beta: Matrices of the fitted SHIR estimator. The m-th column contains the fitted coefficient beta of the m-th site, and the first row contains the fitted intercepts.